requirements.txt
anyio==4.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.3.0
fonttools==4.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
packaging==25.0
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
typing-extensions==4.13.2
urllib3==2.2.3
zipp==3.20.2
apscheduler==3.10.4
